config for trialverse datasource

+++ src/main/java/org/drugis/addis/trialverse/Trialverse.java
@@ -11,7 +11,7 @@ import javax.persistence.Table;
 @Table(name = "namespaces")
 public class Trialverse {
   @Id
-  private Integer id;
+  private Long id;
 
   private String name;
 
@@ -20,17 +20,17 @@ public class Trialverse {
   public Trialverse() {
   }
 
-  public Trialverse(Integer id, String name, String description) {
+  public Trialverse(Long id, String name, String description) {
     this.id = id;
     this.name = name;
     this.description = description;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
